Completed bonus (juggling multiple files)
As the bonus allows only a single static variable, I used a linked list.
To stay in accordance with the norm, I can only have 5 functions and a .c
file. This constraint made it difficult for me to free up nodes, so for
now they remain there as a leak.

When adding to my libft, I'll most likely create some helper files to
be able to clean up, as I'm not satisfied with the code as it is.